requirements.txt
selenium==4.15.2
googlesearch-python==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
streamlit==1.29.0
pandas==2.1.0
scikit-learn==1.3.0
nltk==3.8.1
newspaper3k==0.2.8
tldextract==3.4.4
readability-lxml==0.8.1
